live_pipeline/transform/requirements.txt
pandas
ipykernel 
boto3
python-dotenv